requirements.txt
albucore==0.0.24
albumentations==2.0.8
annotated-types==0.7.0
captum==0.8.0
certifi==2025.6.15
charset-normalizer==3.4.2
choreographer==1.0.9
contourpy==1.3.2
cycler==0.12.1
et_xmlfile==2.0.0
filelock==3.13.1
fonttools==4.58.4
fsspec==2024.6.1
grad-cam==1.5.5
hf-xet==1.1.5
huggingface-hub==0.34.4
idna==3.10
imageio==2.37.0
Jinja2==3.1.4
joblib==1.5.1
kaleido==1.0.0
kiwisolver==1.4.8
lazy_loader==0.4
logistro==1.1.0
MarkupSafe==2.1.5
matplotlib==3.10.3
mpmath==1.3.0
narwhals==1.48.0
networkx==3.3
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.1.105
nvidia-nvtx-cu12==12.1.105
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
orjson==3.11.0
packaging==25.0
pandas==2.3.0
pillow==11.0.0
plotly==6.2.0
pydantic==2.11.7
pydantic_core==2.33.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
regex==2025.7.34
requests==2.32.4
safetensors==0.5.3
scikit-image==0.25.2
scikit-learn==1.7.0
scipy==1.16.0
seaborn==0.13.2
setuptools==70.2.0
simplejson==3.20.1
simsimd==6.4.9
six==1.17.0
stringzilla==3.12.5
sympy==1.13.1
threadpoolctl==3.6.0
tifffile==2025.6.11
timm==1.0.19
tokenizers==0.21.4
torch==2.5.1+cu121
torchaudio==2.5.1+cu121
torchvision==0.20.1+cu121
tqdm==4.67.1
transformers @ git+https://github.com/huggingface/transformers.git@52c6c1bb6e27ca87c4faede34a4c2a7404c17c4d
triton==3.1.0
ttach==0.0.3
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.5.0
